refactor: 모달 UI 리팩토링 [#front-26]
